minor error handling improvements and style adjustments

+++ exec.c
@@ -1,13 +1,12 @@
 /* Copyright 2013 Oliver Giles
- * This file is part of Wemed. Wemed is licensed under the 
+ * This file is part of Wemed. Wemed is licensed under the
  * GNU GPL version 3. See LICENSE or <http://www.gnu.org/licenses/>
  * for more information */
-#include <unistd.h>
 #include <sys/wait.h>
 #include <fcntl.h>
 #include "exec.h"
 
-int exec_get(char* buffer, int nbuf, const char* file, const char* const* args) {
+ssize_t exec_get(char* buffer, size_t nbuf, const char* file, const char* const* args) {
 	int pipes[2];
 
 	pipe(pipes);
@@ -23,12 +22,14 @@ int exec_get(char* buffer, int nbuf, const char* file, const char* const* args) 
 		_exit(0);
 	}
 	waitpid(pid, 0, 0);
-	int n = read(pipes[0], buffer, nbuf);
-	if(n < 0) return n;
+	ssize_t n = read(pipes[0], buffer, nbuf);
+	if(n < 0)
+		return (int) n;
 	buffer[nbuf] = '\0';
 
 	close(pipes[0]);
 	close(pipes[1]);
+
 	return n;
 }
